pyproject.toml
[project]
name = "100-days-of-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.133.0",
    "langchain==0.3.27",
    "langchain-community==0.3.19",
    "langchain-openai==0.2.14",
    "langchain-qdrant==0.2.0",
    "pydantic>=2.12.5",
    "pypdf>=6.7.4",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.17.0",
    "uvicorn>=0.41.0",
]
